pyproject.toml
[project]
name = "super-worker"
version = "0.1.0"
description = "Claude Code Instance Manager for Git Worktrees"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "textual>=0.89.0,<1.0",
    "libtmux>=0.37.0",
    "gitpython>=3.1.0",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23,<1.3",
    "pytest-timeout>=2.0",
]

[project.urls]
Repository = "https://github.com/okeidar/super-worker"

[project.scripts]
sw = "super_worker.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
timeout = 30
